requirements.txt
feedparser~=6.0.2
Flask~=1.1.2
python-dateutil~=2.8.1
Click==7.0
cycler==0.10.0
SQLAlchemy==1.2.14